Added lighting stuff to camera's capture
Lighting is based on taking rgb -> hsv and varying v to be dot product of plane's normal and the lighting vector. Right it *feels* wonky, but I can't figure out what I'm doing wrong or if there even is anything wrong. Strictly speaking the dot product thing *is* working, but IDK if the outcome is because that's just what it is or if I'm using the dot product wrong. Either way, problem for a future me.